Give libraries unique name values in library.properties
Library Manager has a problem where libraries are shown as always updatable if the hardware package of the selected board contains a library with the same name as a library bundled with the Arduino IDE and there is a newer version of the library available that the version of the library in the hardware package. This issue currently occurs for MegaCore boards with the SD library and with Ethernet and Servo libraries in Arduino IDE ca 1.6.8. The backwards compatible workaround is to give the libraries unique name values in library.properties.