Restrict CI backwards compatibility testing to "landmark" IDE versions
Previously, the full range of usable IDE versions was tested for backwards compatibility during the Travis CI build. Attempting to test the full range of IDE versions caused the job length to exceed Travis CI's maximum allowed job duration. For this reason, the IDE version range was split in two, and a job created for each half of the range. This resulted in a significant increase in the number of jobs in the CI build. With the addition of the AT90CAN boards, the job count exceeded Travis CI's maximum of 200. The solution was to restrict backwards compatibility testing to significant IDE versions, which should still provide reasonable backwards compatibility testing coverage while keeping the job count under the maximum (with room to grow) and reduce the build duration.